requirements.txt
pandas==2.1.4
numpy==1.26.4
pytz==2023.3.post1
tqdm==4.67.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
tinkoff==0.1.1
Telethon==1.36.0